requirements.txt
docker==4.3.1
python-dotenv==0.14.0
python-telegram-bot==13.0